% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_mortality.R
\name{annual_mortality}
\alias{annual_mortality}
\title{Annual Mortality}
\usage{
annual_mortality(inventory, per_species = FALSE)
}
\arguments{
\item{inventory}{A \code{data.table} or \code{data.frame} with the inventory data. 
It most contain four columns labeled: 'block', 'plot', 'species', and 'dead'.}

\item{per_species}{Logic. If \code{TRUE}, it estimate the mortality per species
whitin plots.}
}
\value{
It returns a \code{data.table} with the estimations of morality per plot.
}
\description{
It computes the annual mortality on plots since its established.
}
\author{
J. Antonio Guzmán Q.
}
