% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_layout.R
\name{experimental_layout}
\alias{experimental_layout}
\title{Experimental layout}
\usage{
experimental_layout(design, outfile = NULL)
}
\arguments{
\item{design}{A \code{data.frame} or \code{data.table} with the description
of the experimental layout. It most contain columns with the labels: "Block",
"plot", "species", "row", and "column".}

\item{output}{A path and name of the file to export. If \code{NULL}, it will
export a pdf to current working directory under \code{layout.pdf}.}
}
\value{
A \code{PDF} with describing the distribution of individuals within 
each plot.
}
\description{
It creates a pdf of the experiment layout of individuals within 
plots. Useful as a field guide.
}
