% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_edge.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove plot edge}
\usage{
remove_edge(inventory)
}
\arguments{
\item{inventory}{A \code{data.table} or \code{data.frame}. It most contain 
three columns labeled: 'plot', 'row', and 'columns'.}
}
\value{
A \code{data.table} without at the edge line of the plots.
}
\description{
Remove the border line of trees planted within the plots to 
avoid edge effects in future analyses. The function first identify the edge line
for each plot based on rows and columns, than it remove edge trees.
}
