% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_plots.R
\name{split_plots}
\alias{split_plots}
\title{Split large plots}
\usage{
split_plots(inventory)
}
\arguments{
\item{inventory_large}{A \code{data.table} or \code{data.frame} with the inventory data.
It most contain three columns labeled: 'plot', 'row', and 'column'.
Only use plots that require splitting into subplots.}
}
\value{

}
\description{
It split large plots into four subplots.
}
\details{
The function assume that all the plots contain the same amount of rows
and columns of trees planted. It also assume that rows and columns are numbered
continuously from 0 to the max number of rows and columns.
}
