% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Pyhlogenetic or Functional Diversity}
\usage{
diversity(inventory, tree = NULL, weight = c("abundance", "n_indiviuals"))
}
\arguments{
\item{inventory}{A \code{data.table} or \code{data.frame} with the inventory data.
It most contain three columns: 'plot', 'species', and 'values'; in that order.
The last column should be the metric used to weight diversity estimations
(e.g., DBH, above ground biomass, or volume).}

\item{tree}{A \code{phylo} tree object,  or a phylogenetic covariance matrix.}
}
\value{
It returns a \code{data.table} with metrics of diversity.
}
\description{
Derive diversity metrics from phylogenetic trees or cluster trees
of plant traits. This is a function that encapsulate functions from \code{picante}.
}
\details{
The function first summarize forest inventories, then it calculate metrics
of species richness (SR), phylogenetic species variability (PSV),
phylogenetic species richness (PSR), phylogenetic species evenness (PSE),
phylogenetic species clustering (PSC), and Faith's phylogenetic diversity. Similar
metrics can be derived on functional information, but using an \code{hclust} tree
of similarity of traits among species. If you use this function in your work,
please also acknowledge \code{picante} package.
}
\examples{
Create file for compile results
}
\author{
J. Antonio Guzmán Q.
}
